Add version invalidation logic and test
