+++ postgres/tests/test_integration.py
@@ -7,41 +7,11 @@ import psycopg2
 import pytest
 
 from datadog_checks.postgres import PostgreSql
+from .utils import requires_over_10
+from .common import check_bgw_metrics, check_common_metrics
 
-from .common import DB_NAME, HOST, PORT
+from .common import DB_NAME, HOST, PORT, POSTGRES_VERSION
 
-COMMON_METRICS = [
-    'postgresql.before_xid_wraparound',
-    'postgresql.connections',
-    'postgresql.commits',
-    'postgresql.rollbacks',
-    'postgresql.disk_read',
-    'postgresql.buffer_hit',
-    'postgresql.rows_returned',
-    'postgresql.rows_fetched',
-    'postgresql.rows_inserted',
-    'postgresql.rows_updated',
-    'postgresql.rows_deleted',
-    'postgresql.database_size',
-    'postgresql.deadlocks',
-    'postgresql.temp_bytes',
-    'postgresql.temp_files',
-]
-
-COMMON_BGW_METRICS = [
-    'postgresql.bgwriter.checkpoints_timed',
-    'postgresql.bgwriter.checkpoints_requested',
-    'postgresql.bgwriter.buffers_checkpoint',
-    'postgresql.bgwriter.buffers_clean',
-    'postgresql.bgwriter.maxwritten_clean',
-    'postgresql.bgwriter.buffers_backend',
-    'postgresql.bgwriter.buffers_alloc',
-    'postgresql.bgwriter.buffers_backend_fsync',
-    'postgresql.bgwriter.write_time',
-    'postgresql.bgwriter.sync_time',
-]
-
-COMMON_BGW_METRICS_PG_ABOVE_94 = ['postgresql.archiver.archived_count', 'postgresql.archiver.failed_count']
 
 CONNECTION_METRICS = ['postgresql.max_connections', 'postgresql.percent_usage_connections']
 
@@ -51,20 +21,6 @@ ACTIVITY_METRICS = [
     'postgresql.active_queries',
     'postgresql.waiting_queries',
 ]
-
-
-def check_common_metrics(aggregator, expected_tags, count=1):
-    for name in COMMON_METRICS:
-        aggregator.assert_metric(name, count=count, tags=expected_tags)
-
-
-def check_bgw_metrics(aggregator, expected_tags):
-    for name in COMMON_BGW_METRICS:
-        aggregator.assert_metric(name, count=1, tags=expected_tags)
-
-    if float(os.environ['POSTGRES_VERSION']) >= 9.4:
-        for name in COMMON_BGW_METRICS_PG_ABOVE_94:
-            aggregator.assert_metric(name, count=1, tags=expected_tags)
 
 
 @pytest.mark.integration
@@ -156,3 +112,18 @@ def test_activity_metrics(aggregator, check, pg_instance):
     expected_tags = pg_instance['tags'] + ['server:{}'.format(HOST), 'port:{}'.format(PORT), 'db:datadog_test']
     for name in ACTIVITY_METRICS:
         aggregator.assert_metric(name, count=1, tags=expected_tags)
+
+
+@requires_over_10
+@pytest.mark.integration
+@pytest.mark.usefixtures('dd_environment')
+def test_wrong_version(aggregator, check, pg_instance):
+    # Enforce to cache wrong version
+    db_key = ('localhost', 5432, 'datadog_test')
+    check.versions[db_key] = [9, 6, 0]
+
+    check.check(pg_instance)
+    assert db_key not in check.versions  # version invalidated
+
+    check.check(pg_instance)
+    assert check.versions[db_key][0] == int(POSTGRES_VERSION)
